Rework DAOSFS file I/O to use array objects
This work has landed in the main daos repo, so we should take
advantage of it to do better I/O.

Also:
  * Change all file size/offset values to int64 (to match other Go code)

+++ cmd/daos-fuse/node.go
@@ -24,7 +24,7 @@ func (n *Node) Attr(ctx context.Context, attr *fuse.Attr) error {
 	}
 
 	attr.Inode = da.Inode
-	attr.Size = da.Size
+	attr.Size = uint64(da.Size)
 	//attr.Blocks = da.Blocks
 	//attr.Atime = da.Atime
 	attr.Mtime = da.Mtime
